Move the components into subfolders to clean up the directory a bit
