--- src/Pages.js
@@ -1,13 +1,0 @@
-import AboutMeRouteInfo from './AboutMeRouteInfo';
-import ContactMeRouteInfo from './ContactMeRouteInfo';
-import ProjectsRouteInfo from './ProjectsRouteInfo';
-import ProjectDetailsRouteInfo from './ProjectDetailsRouteInfo';
-
-const Pages = [
-  AboutMeRouteInfo,
-  ProjectDetailsRouteInfo,
-  ProjectsRouteInfo,
-  ContactMeRouteInfo,
-];
-
-export default Pages;
