+++ src/routes.js
@@ -1,0 +1,13 @@
+import AboutMeRouteInfo from './AboutMe/AboutMeRouteInfo';
+import ContactMeRouteInfo from './ContactMe/ContactMeRouteInfo';
+import ProjectDetailsRouteInfo from './Projects/ProjectDetailsRouteInfo';
+import ProjectsRouteInfo from './Projects/ProjectsRouteInfo';
+
+const Pages = [
+  AboutMeRouteInfo,
+  ProjectDetailsRouteInfo,
+  ProjectsRouteInfo,
+  ContactMeRouteInfo,
+];
+
+export default Pages;
